Update eb_av1_wiener_convolve_add_src AVX2 and add AVX512
Speed gains of new AVX2 vs old AVX2 with size 64x64:
Tap 7: 10%
Tap 5: 27%
Tap 3: 95%

Speed gains of AVX512 vs new AVX2 with size 64x64:
Tap 7: 50%
Tap 5: 50%
Tap 3: 60%

+++ Source/Lib/Common/ASM_SSE2/synonyms.h
@@ -80,25 +80,25 @@ static INLINE __m128i load_u16_4x2_sse2(const uint16_t *const src,
 }
 
 SIMD_INLINE void store_u8_4x2_sse2(const __m128i src, uint8_t *const dst,
-    const int32_t stride) {
+    const ptrdiff_t stride) {
     xx_storel_32(dst, src);
     *(int32_t *)(dst + stride) = _mm_extract_epi32(src, 1);
 }
 
 SIMD_INLINE void store_u16_2x2_sse2(const __m128i src, uint16_t *const dst,
-    const int32_t stride) {
+    const ptrdiff_t stride) {
     xx_storel_32(dst, src);
     *(int32_t *)(dst + stride) = _mm_extract_epi32(src, 1);
 }
 
 SIMD_INLINE void store_s16_4x2_sse2(const __m128i src, int16_t *const dst,
-    const int32_t stride) {
+    const ptrdiff_t stride) {
     _mm_storel_epi64((__m128i *)dst, src);
     _mm_storeh_epi64((__m128i *)(dst + stride), src);
 }
 
 SIMD_INLINE void store_u16_4x2_sse2(const __m128i src, uint16_t *const dst,
-    const int32_t stride) {
+    const ptrdiff_t stride) {
     _mm_storel_epi64((__m128i *)dst, src);
     _mm_storeh_epi64((__m128i *)(dst + stride), src);
 }
